+++ src/components/MeetingExitMenu/MeetingExitBtn/MeetingExitBtn.stories.tsx
