+++ src/components/MeetingExitMenu/MeetingExitBtn/MeetingExitBtn.styled.ts
